Bug fix on damage text of hero

+++ core/src/com/group12/cookiesrising/GameObjects/Hero.java
@@ -97,7 +97,7 @@ public abstract class Hero extends AbstractGameObject implements Hittable,Health
         this.attackPoint = 1.0D;
         this.healthPoint = 10.0D;
         this.maxhealthPoint = 10D;
-        dmgText = attackPoint+"";
+        dmgText = (int)attackPoint+"";
         speed = 5;
     }
